+++ src/components/styles/Text.css
@@ -1,6 +1,6 @@
 .text-base {
   font-size: 25px;
-  font-family: Roboto;
+  font-family: Roboto-Regular;
   color: #424C50;
 }
